+++ projects/angular-sdk/src/apis/hopper-cloud-airline/v1/model/cfarOfferCustomer.ts
@@ -52,4 +52,5 @@ export interface CfarOfferCustomer {
     createdDateTime: Date;
     itinerary: CfarItinerary;
     offerDescription: MapStringArray;
+    termsConditionsUrl: string;
 }